requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.7.3
dnspython==2.4.2
tldextract==3.4.0
